refactor: Isolate card secure logic and centralize handlers for future
The upcoming hosted iframe tokenization process will use some of the
existing logic (display of errors, appending inputs to the form) so it
makes sense to keep some of this logic in place. This refactor makes
some notable changes to the file structure with the plan of making later
updates more straightforward and concise.
