+++ javascript/src/checkout-card-secure-tokenizer.ts
@@ -1,11 +1,4 @@
-interface ICardConnectResponse {
-  action: string;
-  data: string;
-}
-
-interface JQueryGlobal extends JQueryStatic {
-    payment: any;
-}
+import { JQueryGlobal, ICardConnectResponse } from './interfaces';
 
 export default class WoocommereCardConnect {
